ONB-58 refactor: GameSortType import 변경
