+++ domain/src/main/java/com/yapp/bol/domain/usecase/group/GetGroupJoinedUseCase.kt
@@ -1,10 +1,10 @@
 package com.yapp.bol.domain.usecase.group
 
 import com.yapp.bol.domain.model.GameItem
+import com.yapp.bol.domain.model.GameSortType
 import com.yapp.bol.domain.model.GetGroupJoinedItem
 import com.yapp.bol.domain.model.GroupDetailItem
 import com.yapp.bol.domain.repository.GameRepository
-import com.yapp.bol.domain.repository.GameSortType
 import com.yapp.bol.domain.repository.GroupRepository
 import com.yapp.bol.domain.repository.UserRepository
 import com.yapp.bol.domain.utils.doWork
